+++ hype-gcs/pom.xml
@@ -11,6 +11,12 @@
 
   <dependencies>
     <dependency>
+      <groupId>com.spotify</groupId>
+      <artifactId>hype-common</artifactId>
+      <version>0.0.18-SNAPSHOT</version>
+    </dependency>
+
+    <dependency>
       <groupId>com.google.cloud</groupId>
       <artifactId>google-cloud-storage</artifactId>
     </dependency>
